removed image formats, moved media overlay code into seperate file
