+++ src/Sulu/Bundle/MediaBundle/DependencyInjection/SuluMediaExtension.php
@@ -55,7 +55,6 @@ class SuluMediaExtension extends Extension implements PrependExtensionInterface
                     'image_format_files' => [
                         '%kernel.root_dir%/config/image-formats.xml',
                         __DIR__ . '/../Resources/config/image-formats.xml',
-                        __DIR__ . '/../Resources/config/image-formats-v2.xml',
                     ],
                     'search' => ['enabled' => $container->hasExtension('massive_search')],
                 ]
