+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaSelection/mediaSelectionOverlay.scss
@@ -1,6 +1,6 @@
 @import 'sulu-admin-bundle/containers/Application/variables.scss';
 
-.media-overlay {
+.overlay {
     width: 1170px;
     height: 650px;
     overflow-y: auto;
