+++ src/Sulu/Bundle/MediaBundle/Resources/translations/admin.de.json
@@ -1,7 +1,7 @@
 {
     "sulu_media.all_media": "Alle Medien",
     "sulu_media.copy_url": "URL kopieren",
-    "sulu_media.download_masterfile": "Download master file",
+    "sulu_media.download_masterfile": "Original herunterladen",
     "sulu_media.select_media": "Medien auswählen",
     "sulu_media.reset_selection": "Auswahl zurücksetzen"
 }
